+++ package.json
@@ -14,6 +14,7 @@
     "@zeit/next-css": "^1.0.1",
     "dotenv": "^8.0.0",
     "next": "^9.0.3",
+    "next-page-transitions": "^1.0.0-beta.2",
     "react": "^16.9.0",
     "react-dom": "^16.12.0"
   },
