+++ pages/_app.tsx
@@ -1,7 +1,72 @@
-import React from "react";
+import React, { MouseEvent } from "react";
+import Link from "next/link";
+import { PageTransition } from "next-page-transitions";
+import "../styles/index.css";
 
 const App = ({ Component, pageProps }) => {
-  return <Component {...pageProps} />;
+  let bgRef;
+
+  const hoverColor = (e: MouseEvent) => {
+    const x: number = window.innerWidth;
+    const y: number = window.innerHeight;
+    const mX: number = e.clientX;
+    const mY: number = e.clientY;
+
+    if (!bgRef) {
+      return;
+    }
+
+    bgRef.style.background = `linear-gradient(130deg, rgb(${(200 * mX) /
+      x},${(200 * mY) / y},200), rgb(${(200 * mY) / y},120,${(200 * x) / mX})), 
+      linear-gradient(210deg, rgb(${(200 * mX) / x},${(200 * mY) / y},${(200 *
+      mX) /
+      x}), rgb(160,${(200 * mY) / y},${(200 * mX) / x})),
+        linear-gradient(330deg, rgb(80,${(200 * y) / mY},${(200 * mX) /
+      x}), rgb(100,${(200 * mY) / y},${(200 * mX) / x}))`;
+
+    console.log("mouse x", mX / x, "mouse y", mY / y);
+  };
+
+  return (
+    <>
+      <div className="slowFade" onMouseMove={hoverColor}>
+        <div ref={node => (bgRef = node)} className="App vh vw fix bg-blue" />
+        <div className="vh vw fix noise border" />
+
+        <div className="top right abs my1 mx2 pt05 white o-4 z10 pointer">
+          <span className="mr1">
+            <Link href="/index">Projects</Link>
+          </span>
+          <Link href="/about">About</Link>
+        </div>
+
+        <div className="aic jcc f fw p1 yview">
+          <div className="container-m px2 fa ac o-4 white">
+            <PageTransition timeout={500} classNames="page-transition">
+              <Component {...pageProps} />
+            </PageTransition>
+          </div>
+        </div>
+      </div>
+
+      <style jsx global>{`
+        .page-transition-enter {
+          opacity: 0;
+        }
+        .page-transition-enter-active {
+          opacity: 1;
+          transition: opacity 300ms;
+        }
+        .page-transition-exit {
+          opacity: 1;
+        }
+        .page-transition-exit-active {
+          opacity: 0;
+          transition: opacity 300ms;
+        }
+      `}</style>
+    </>
+  );
 };
 
 export default App;
