+++ pages/index.tsx
@@ -1,64 +1,58 @@
-import React, { useState, MouseEvent } from "react";
-import About from "../components/about";
-import Home from "../components/home";
-import Head from "next/head";
-import "../styles/index.css";
+import React from "react";
+import Fader from "../components/fader";
 import { NextPage } from "next";
 
-const App: NextPage = () => {
-  let bgRef;
-  const [fade, setFade] = useState<"fadeIn" | "fadeOut" | "none">("none");
-
-  const hoverColor = (e: MouseEvent) => {
-    const x: number = window.innerWidth;
-    const y: number = window.innerHeight;
-    const mX: number = e.clientX;
-    const mY: number = e.clientY;
-
-    if (!bgRef) {
-      return;
-    }
-
-    bgRef.style.background = `linear-gradient(130deg, rgb(${(200 * mX) /
-      x},${(200 * mY) / y},200), rgb(${(200 * mY) / y},120,${(200 * x) / mX})), 
-      linear-gradient(210deg, rgb(${(200 * mX) / x},${(200 * mY) / y},${(200 *
-      mX) /
-      x}), rgb(160,${(200 * mY) / y},${(200 * mX) / x})),
-        linear-gradient(330deg, rgb(80,${(200 * y) / mY},${(200 * mX) /
-      x}), rgb(100,${(200 * mY) / y},${(200 * mX) / x}))`;
-
-    console.log("mouse x", mX / x, "mouse y", mY / y);
-  };
-
-  const toggleHome = () => {
-    if (fade === "fadeIn") {
-      setFade("fadeOut");
-      return setTimeout(() => setFade("none"), 500);
-    } else {
-      setFade("fadeIn");
-    }
-  };
-
-  return (
-    <div className="slowFade" onMouseMove={hoverColor}>
-      <div ref={node => (bgRef = node)} className="App vh vw fix bg-blue" />
-      <div className="vh vw fix noise border" />
-
-      <div className="top right abs my1 mx2 pt05 white o-4 z10 pointer">
-        <span className="mr1" onClick={toggleHome}>
-          Projects
-        </span>
-        <span onClick={toggleHome}>About</span>
-      </div>
-
-      <div className="aic jcc f fw p1 yview">
-        <div className="container-m px2 fa ac o-4 white">
-          <About style={fade} toggle={toggleHome} />
-          <Home />
-        </div>
-      </div>
+const Home: NextPage = () => (
+  <div className="fadeIn">
+    <div className="__headline">
+      <h1>
+        Max Smouha is a developer <Fader />
+      </h1>
     </div>
-  );
-};
 
-export default App;
+    <div className="my2">
+      <p className="h6 kern">RECENT PROJECTS</p>
+    </div>
+
+    <div className="f fw mt1">
+      {links.map((link, i) => (
+        <div className="col-25 px075 my1" key={i}>
+          <a href={link.url} target="_blank" rel="noopener noreferrer">
+            <h5>{link.name}</h5>
+          </a>
+          <p className="mt05">{link.desc}</p>
+        </div>
+      ))}
+    </div>
+
+    <div className="x left ac mt1 mb2">
+      <a href="mailto:maxsmouha@gmail.com">email</a>{" "}
+      <a href="https://github.com/squamgulia">git</a>{" "}
+    </div>
+  </div>
+);
+
+export default Home;
+
+const links: Array<{
+  name: string;
+  url: string;
+  desc: string;
+}> = [
+  {
+    name: "TriBeCa Art Walk",
+    url: "https://www.tribecaarts.org",
+    desc: "react, mapbox"
+  },
+  { name: "Fridge Freud", url: "http://freud.fun", desc: "react, node" },
+  {
+    name: "Piano Friend",
+    url: "http://pianofriend.cool/",
+    desc: "vanilla js"
+  },
+  {
+    name: "Net Effective Rent",
+    url: "https://squamuglia.github.io/neteffect/",
+    desc: "vanilla js"
+  }
+];
