+++ styles/components/animations.css
@@ -70,3 +70,18 @@
     background-position: 0% 50%;
   }
 }
+
+.page-transition-enter {
+  opacity: 0;
+}
+.page-transition-enter-active {
+  opacity: 1;
+  transition: opacity 500ms;
+}
+.page-transition-exit {
+  opacity: 1;
+}
+.page-transition-exit-active {
+  opacity: 0;
+  transition: opacity 500ms;
+}
